Move EventType into eventmodel.py
